+++ _define.php
@@ -12,9 +12,6 @@
  */
 declare(strict_types=1);
 
-// can not use namespace as it is included inside a class method,
-// and also can be included in third party plugins class methods.
-
 if (!defined('DC_RC_PATH')) {
     return null;
 }
@@ -25,7 +22,7 @@ $this->registerModule(
     'Osku and contributors',
     '0.3.1',
     [
-        'requires'    => [['core', '2.24']],
+        'requires'    => [['core', '2.26']],
         'permissions' => null,
         'type'        => 'plugin',
         'support'     => 'https://github.com/JcDenis/' . basename(__DIR__),