+++ src/My.php
@@ -1,0 +1,38 @@
+<?php
+/**
+ * @brief testMail, a plugin for Dotclear 2
+ *
+ * @package Dotclear
+ * @subpackage Plugin
+ *
+ * @author Osku and contributors
+ *
+ * @copyright Jean-Christian Denis
+ * @copyright GPL-2.0 https://www.gnu.org/licenses/gpl-2.0.html
+ */
+declare(strict_types=1);
+
+namespace Dotclear\Plugin\testMail;
+
+use dcCore;
+
+class My
+{
+    public const X_MAILER = 'Dotclear';
+
+    /**
+     * This module id
+     */
+    public static function id(): string
+    {
+        return basename(dirname(__DIR__));
+    }
+
+    /**
+     * This module name
+     */
+    public static function name(): string
+    {
+        return __((string) dcCore::app()->plugins->moduleInfo(self::id(), 'name'));
+    }
+}